requirements.txt
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
certifi==2024.12.14
click==8.1.8
fastapi==0.115.6
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
logzero==1.7.0
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
pydantic==2.10.4
pydantic_core==2.27.2
pytest==8.3.4
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
tzlocal==5.2
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.3
websockets==14.1
